requirements.txt
fastapi
uvicorn[standard]
pydantic
sentence-transformers
pinecone
google-generativeai
numpy
faiss-cpu
tqdm